rename exportMap hook to useExportMap
