+++ src/components/MlCreatePdfForm/MlCreatePdfForm.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import PdfPreview from './lib/PdfPreview';
 import PdfForm from './lib/PdfForm';
 import { PdfContextProvider } from './lib/PdfContext';
-import { createPdfResolverParams } from '../../hooks/exportMap/lib';
+import { createPdfResolverParams } from '../../hooks/useExportMap/lib';
 import { SxProps } from '@mui/material';
 
 interface MlCreatePdfFormProps {
